+++ TicketingScreenDesigner/DataAccessLayer/ScreenOperations.cs
@@ -8,19 +8,76 @@ namespace DataAccessLayer {
 	/// <summary>
 	/// Contains methods for retrieving and manipulating screen information.
 	/// </summary>
-	public static class ScreenOperations {
-		private static readonly SqlConnection connection = DBUtils.CreateConnection();
-		
-		private static class ActiveScreenController {
-			public static int? GetActiveScreenId(string bankName) {
-				int? ret = -1;
+	public class ScreenOperations {
+		public static readonly ScreenOperations Instance = new();
+		private SqlConnection? connection;
 
+		private ScreenOperations() {
+			try {
+				connection = null;
+				connection = DBUtils.CreateConnection();
+			}
+			catch (Exception ex) {
+				Console.Error.WriteLine($"Failed to establish connection. Message: {ex.Message}");
+			}
+		}
+
+		/// <summary>
+		/// Attempts to establish the connection again.
+		/// </summary>
+		/// <returns><c>true</c> if the connection was established successfully, and <c>false</c> otherwise.</returns>
+		public bool ReinitializeConnection() {
+			try {
+				connection = DBUtils.CreateConnection();
+				return true;
+			}
+			catch {
+				return false;
+			}
+		}
+
+		/// <summary>
+		/// Verifies that the database connection is established correctly.
+		/// </summary>
+		/// <returns><c>true</c> if the connection has been established properly, and <c>false</c> otherwise.</returns>
+		public bool VerifyConnection() {
+			if (connection is null)
+				return false;
+
+			try {
+				connection.Open();
+				var command = new SqlCommand($"SELECT 1 FROM {ScreensConstants.TABLE_NAME};", connection);
+
+				var result = command.ExecuteScalar();
+				if (result is null)
+					return false;
+
+				return (int) result == 1;
+			}
+			catch (Exception ex) {
+				Console.Error.WriteLine($"Failed to establish connection. Message: {ex.Message}");
+				return false;
+			}
+			finally {
+				connection?.Close();
+			}
+		}
+
+		private class ActiveScreenController {
+			public static readonly ActiveScreenController Instance = new();
+
+			public int? GetActiveScreenId(string bankName) {
 				try {
+					if (ScreenOperations.Instance.connection is null)
+						return null;
+
+					int? ret = -1;
+
 					string query = $"SELECT {ScreensConstants.SCREEN_ID} FROM {ScreensConstants.TABLE_NAME} WHERE {ScreensConstants.BANK_NAME} = @bankName AND {ScreensConstants.IS_ACTIVE} = 1;";
-					var command = new SqlCommand(query, connection);
+					var command = new SqlCommand(query, ScreenOperations.Instance.connection);
 					command.Parameters.AddWithValue("@bankName", bankName);
 
-					connection.Open();
+					ScreenOperations.Instance.connection.Open();
 
 					var reader = command.ExecuteReader();
 
@@ -36,17 +93,17 @@ namespace DataAccessLayer {
 					ExceptionHelper.HandleGeneralException(ex);
 				}
 				finally {
-					connection.Close();
+					ScreenOperations.Instance.connection?.Close();
 				}
 
 				return null;
 			}
 
-			public static bool DeactivateScreen(string bankName, int screenId) {
+			public bool DeactivateScreen(string bankName, int screenId) {
 				return SetIsActive(bankName, screenId, false);
 			}
 
-			public static bool ActivateScreen(string bankName, int screenId) {
+			public bool ActivateScreen(string bankName, int screenId) {
 				int? currentlyActiveScreenId = GetActiveScreenId(bankName);
 
 				if (currentlyActiveScreenId is null)
@@ -59,15 +116,18 @@ namespace DataAccessLayer {
 				return SetIsActive(bankName, screenId, true);
 			}
 
-			private static bool SetIsActive(string bankName, int screenId, bool active) {
+			private bool SetIsActive(string bankName, int screenId, bool active) {
 				try {
+					if (ScreenOperations.Instance.connection is null)
+						return false;
+
 					string query = $"UPDATE {ScreensConstants.TABLE_NAME} SET {ScreensConstants.IS_ACTIVE} = @isActive WHERE {ScreensConstants.BANK_NAME} = @bankName AND {ScreensConstants.SCREEN_ID} = @screenId";
-					var command = new SqlCommand(query, connection);
+					var command = new SqlCommand(query, ScreenOperations.Instance.connection);
 					command.Parameters.AddWithValue("@isActive", active ? 1 : 0);
 					command.Parameters.AddWithValue("@bankName", bankName);
 					command.Parameters.AddWithValue("@screenId", screenId);
 
-					connection.Open();
+					ScreenOperations.Instance.connection.Open();
 
 					return command.ExecuteNonQuery() == 1;
 				}
@@ -78,7 +138,7 @@ namespace DataAccessLayer {
 					ExceptionHelper.HandleGeneralException(ex);
 				}
 				finally {
-					connection.Close();
+					ScreenOperations.Instance.connection?.Close();
 				}
 
 				return false;
@@ -90,8 +150,11 @@ namespace DataAccessLayer {
 		/// </summary>
 		/// <param name="screen">The screen to be added to the database</param>
 		/// <returns>The ID of the screen that was added. If the operation fails, <c>null</c> is returned.</returns>
-		public static int? AddScreen(TicketingScreen screen) {
+		public int? AddScreen(TicketingScreen screen) {
 			try {
+				if (connection is null)
+					return null;
+
 				if (screen.IsActive) {
 					int? activeScreenId = GetActiveScreenId(screen.BankName);
 					if (activeScreenId is null)
@@ -117,7 +180,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return null;
@@ -128,8 +191,11 @@ namespace DataAccessLayer {
 		/// </summary>
 		/// <param name="bankName">The name of the bank.</param>
 		/// <returns>The ID of the active screen. If there is no active screen, <c>-1</c> is returned. If the operation fails, <c>null</c> is returned.</returns>
-		public static int? GetActiveScreenId(string bankName) {
-			return ActiveScreenController.GetActiveScreenId(bankName);
+		public int? GetActiveScreenId(string bankName) {
+			if (connection is null)
+				return null;
+
+			return ActiveScreenController.Instance.GetActiveScreenId(bankName);
 		}
 
 		/// <summary>
@@ -138,8 +204,11 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank which owns the screen.</param>
 		/// <param name="screenId">The ID of the screen.</param>
 		/// <returns><c>true</c> if the operation succeeds, and <c>false</c> if it fails.</returns>
-		public static bool DeactivateScreen(string bankName, int screenId) {
-			return ActiveScreenController.DeactivateScreen(bankName, screenId);
+		public bool DeactivateScreen(string bankName, int screenId) {
+			if (connection is null)
+				return false;
+
+			return ActiveScreenController.Instance.DeactivateScreen(bankName, screenId);
 		}
 
 		/// <summary>
@@ -148,8 +217,11 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank which owns the screen.</param>
 		/// <param name="screenId">The ID of the screen.</param>
 		/// <returns><c>true</c> if the operation succeeds, and <c>false</c> if it fails.</returns>
-		public static bool ActivateScreen(string bankName, int screenId) {
-			return ActiveScreenController.ActivateScreen(bankName, screenId);
+		public bool ActivateScreen(string bankName, int screenId) {
+			if (connection is null)
+				return false;
+
+			return ActiveScreenController.Instance.ActivateScreen(bankName, screenId);
 		}
 
 		/// <summary>
@@ -158,8 +230,11 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank that owns the screen.</param>
 		/// <param name="screenId">The ID of the screen to be deleted</param>
 		/// <returns><c>true</c> if the operation succeeds, and <c>false</c> if it fails.</returns>
-		public static bool DeleteScreen(string bankName, int screenId) {
+		public bool DeleteScreen(string bankName, int screenId) {
 			try {
+				if (connection is null)
+					return false;
+
 				string query = $"DELETE FROM {ScreensConstants.TABLE_NAME} WHERE {ScreensConstants.BANK_NAME} = @bankName AND {ScreensConstants.SCREEN_ID} = @screenId;";
 				var command = new SqlCommand(query, connection);
 				command.Parameters.AddWithValue("@bankName", bankName);
@@ -176,7 +251,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return false;
@@ -188,8 +263,11 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank that owns the screens.</param>
 		/// <param name="screenIds">The IDs of the screen to be deleted</param>
 		/// <returns><c>true</c> if the operation succeeds, and <c>false</c> if it fails.</returns>
-		public static bool DeleteScreens(string bankName, List<int> screenIds) {
+		public bool DeleteScreens(string bankName, List<int> screenIds) {
 			try {
+				if (connection is null)
+					return false;
+
 				var query = new StringBuilder($"DELETE FROM {ScreensConstants.TABLE_NAME} WHERE {ScreensConstants.BANK_NAME} = @bankName AND {ScreensConstants.SCREEN_ID} IN (");
 				var command = new SqlCommand() { Connection = connection };
 				command.Parameters.AddWithValue("@bankName", bankName);
@@ -216,7 +294,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return false;
@@ -229,8 +307,11 @@ namespace DataAccessLayer {
 		/// <param name="screenId">The ID of the screen.</param>
 		/// <param name="newScreen">A <c>TicketingScreen</c> object containing the updated screen information.</param>
 		/// <returns><c>true</c> if the operation succeeds, and <c>false</c> if it fails.</returns>
-		public static bool UpdateScreen(string bankName, int screenId, TicketingScreen newScreen) {
+		public bool UpdateScreen(string bankName, int screenId, TicketingScreen newScreen) {
 			try {
+				if (connection is null)
+					return false;
+
 				int? activeScreenId = GetActiveScreenId(bankName);
 				if (activeScreenId is not null && activeScreenId != screenId) {
 					DeactivateScreen(bankName, (int) activeScreenId);
@@ -254,7 +335,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return false;
@@ -266,8 +347,11 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank that owns the screen.</param>
 		/// <param name="screenId">The ID of the screen.</param>
 		/// <returns><c>true</c> if a matching screen exists, and <c>false</c> if it does not. If the operation fails, <c>null</c> is returned.</returns>
-		public static bool? CheckIfScreenExists(string bankName, int screenId) {
+		public bool? CheckIfScreenExists(string bankName, int screenId) {
 			try {
+				if (connection is null)
+					return null;
+
 				string query = $"SELECT COUNT({ScreensConstants.SCREEN_ID}) FROM {ScreensConstants.TABLE_NAME} WHERE {ScreensConstants.BANK_NAME} = @bankName AND {ScreensConstants.SCREEN_ID} = @screenId;";
 				var command = new SqlCommand(query, connection);
 				command.Parameters.AddWithValue("@bankName", bankName);
@@ -284,7 +368,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return null;
@@ -296,10 +380,13 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank that owns the screen.</param>
 		/// <param name="screenId">The ID of the screen.</param>
 		/// <returns>A <c>TicketingScreen</c> object representing the screen. If the screen does not exist, <EmptyScreen.Value cref="EmptyScreen.Value"/> is returned. If the operation fails, <c>null</c> is returned.</returns>
-		public static TicketingScreen? GetScreenById(string bankName, int screenId) {
-			TicketingScreen? screen = EmptyScreen.Value;
-
+		public TicketingScreen? GetScreenById(string bankName, int screenId) {
 			try {
+				if (connection is null)
+					return null;
+
+				TicketingScreen? screen = EmptyScreen.Value;
+
 				string query = $"SELECT {ScreensConstants.SCREEN_TITLE}, {ScreensConstants.IS_ACTIVE} FROM {ScreensConstants.TABLE_NAME} WHERE {ScreensConstants.BANK_NAME} = @bankName AND {ScreensConstants.SCREEN_ID} = @screenId;";
 				var command = new SqlCommand(query, connection);
 				command.Parameters.AddWithValue("@bankName", bankName);
@@ -324,7 +411,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return null; ;
@@ -336,16 +423,19 @@ namespace DataAccessLayer {
 		/// <param name="bankName">The name of the bank that owns the screen.</param>
 		/// <param name="screenId">The ID of the screen.</param>
 		/// <returns>A list of <c>TicketingButton</c> objects representing the buttons on the screen. If there are no buttons, an empty list is returned. If the operation fails, <c>null</c> is returned.</returns>
-		public static List<TicketingButton>? GetButtons(string bankName, int screenId) {
-			var buttons = new List<TicketingButton>();
-
-			string query = $"SELECT {ButtonsConstants.BUTTON_ID}, {ButtonsConstants.NAME_EN}, {ButtonsConstants.NAME_AR}, {ButtonsConstants.TYPE}, {ButtonsConstants.SERVICE}, {ButtonsConstants.MESSAGE_EN}, " +
-				$"{ButtonsConstants.MESSAGE_AR} FROM {ButtonsConstants.TABLE_NAME} WHERE {ButtonsConstants.BANK_NAME} = @bankName AND {ButtonsConstants.SCREEN_ID} = @screenId";
-			var command = new SqlCommand(query, connection);
-			command.Parameters.AddWithValue("@bankName", bankName);
-			command.Parameters.AddWithValue("@screenId", screenId);
-
+		public List<TicketingButton>? GetButtons(string bankName, int screenId) {
 			try {
+				if (connection is null)
+					return null;
+
+				var buttons = new List<TicketingButton>();
+
+				string query = $"SELECT {ButtonsConstants.BUTTON_ID}, {ButtonsConstants.NAME_EN}, {ButtonsConstants.NAME_AR}, {ButtonsConstants.TYPE}, {ButtonsConstants.SERVICE}, {ButtonsConstants.MESSAGE_EN}, " +
+					$"{ButtonsConstants.MESSAGE_AR} FROM {ButtonsConstants.TABLE_NAME} WHERE {ButtonsConstants.BANK_NAME} = @bankName AND {ButtonsConstants.SCREEN_ID} = @screenId";
+				var command = new SqlCommand(query, connection);
+				command.Parameters.AddWithValue("@bankName", bankName);
+				command.Parameters.AddWithValue("@screenId", screenId);
+
 				connection.Open();
 
 				var reader = command.ExecuteReader();
@@ -380,7 +470,7 @@ namespace DataAccessLayer {
 				ExceptionHelper.HandleGeneralException(ex);
 			}
 			finally {
-				connection.Close();
+				connection?.Close();
 			}
 
 			return null;
